Add files via upload
The data used to train this model was too big to upload to GitHub (todo: upload a script to download the datasets to a local machine). The actual food desert data -- on whether each zip code is a food desert -- was obtained from the USDA, and in this dataset is a binary variable, indicating whether the people a given zip code have a sufficient access to food within a 10 mile radius. In the visualization (".png" file), I wanted to predict how severe the deserts were in particular regions, and because the binary food desert data from the USDA only indicates whether a zip code *is* a food desert or whether it *isn't*, I used ```scipy.signal.convolve``` to get a sense of the severity of food deserts in different regions of the country.